+++ lib/src/connect/connect_button.dart
@@ -116,8 +116,7 @@ class _ZegoAudienceConnectButtonState extends State<ZegoAudienceConnectButton> {
         List<String> errorInvitees,
       ) {
         if (code.isNotEmpty) {
-          widget.connectManager.prebuiltController?.onSeatTakingRequestFailed
-              ?.call();
+          widget.connectManager.config.onSeatTakingRequestFailed?.call();
 
           showDebugToast('Failed to apply for take seat, $code $message');
         } else {