+++ lib/src/connect/connect_manager.dart
@@ -14,6 +14,7 @@ import 'package:zego_uikit_prebuilt_live_audio_room/zego_uikit_prebuilt_live_aud
 
 class ZegoLiveConnectManager {
   ZegoLiveConnectManager({
+    required this.config,
     required this.seatManager,
     required this.prebuiltController,
     required this.innerText,
@@ -22,6 +23,7 @@ class ZegoLiveConnectManager {
     listenStream();
   }
 
+  final ZegoUIKitPrebuiltLiveAudioRoomConfig config;
   final ZegoLiveSeatManager seatManager;
   final ZegoLiveAudioRoomController? prebuiltController;
   final BuildContext Function() contextQuery;
@@ -128,7 +130,7 @@ class ZegoLiveConnectManager {
       if (result.error != null) {
         _audienceIDsInvitedTakeSeatByHost.remove(invitee.id);
 
-        prebuiltController?.onInviteAudienceToTakeSeatFailed?.call();
+        config.onInviteAudienceToTakeSeatFailed?.call();
 
         showDebugToast('Failed to invite take seat, please try again.');
       }
@@ -166,7 +168,7 @@ class ZegoLiveConnectManager {
             List<ZegoUIKitUser>.from(audiencesRequestingTakeSeatNotifier.value)
               ..add(inviter);
 
-        prebuiltController?.onSeatTakingRequested?.call(inviter);
+        config.onSeatTakingRequested?.call(inviter);
       }
     } else {
       if (ZegoInvitationType.inviteToTakeSeat == invitationType) {
@@ -194,7 +196,7 @@ class ZegoLiveConnectManager {
       return;
     }
 
-    prebuiltController?.onHostSeatTakingInviteSent?.call();
+    config.onHostSeatTakingInviteSent?.call();
 
     /// self-cancellation if requesting when host invite you
     ZegoLoggerService.logInfo(
@@ -361,7 +363,7 @@ class ZegoLiveConnectManager {
           List<ZegoUIKitUser>.from(audiencesRequestingTakeSeatNotifier.value)
             ..removeWhere((user) => user.id == inviter.id);
 
-      prebuiltController?.onSeatTakingRequestCanceled?.call(inviter);
+      config.onSeatTakingRequestCanceled?.call(inviter);
     }
 
     /// hide invite take seat dialog
@@ -385,13 +387,13 @@ class ZegoLiveConnectManager {
       _audienceIDsInvitedTakeSeatByHost.remove(invitee.id);
 
       /// host's invite is rejected by audience
-      prebuiltController?.onSeatTakingInviteRejected?.call();
+      config.onSeatTakingInviteRejected?.call();
 
       showDebugToast(
           'Your request to take seat has been refused by ${ZegoUIKit().getUser(invitee.id)?.name ?? ''}');
     } else {
       /// audience's request is rejected by host
-      prebuiltController?.onSeatTakingRequestRejected?.call();
+      config.onSeatTakingRequestRejected?.call();
 
       showDebugToast('Your request to take seat has been refused.');
       updateAudienceConnectState(ConnectState.idle);
@@ -437,7 +439,7 @@ class ZegoLiveConnectManager {
         _audienceIDsInvitedTakeSeatByHost.remove(invitee.id);
       }
     } else {
-      prebuiltController?.onSeatTakingRequestFailed?.call();
+      config.onSeatTakingRequestFailed?.call();
 
       updateAudienceConnectState(ConnectState.idle);
     }