+++ lib/src/live_audio_room.dart
@@ -79,7 +79,7 @@ class _ZegoUIKitPrebuiltLiveAudioRoomState
 
     ZegoUIKit().getZegoUIKitVersion().then((version) {
       ZegoLoggerService.logInfo(
-        'version: zego_uikit_prebuilt_live_audio_room: 2.2.0; $version',
+        'version: zego_uikit_prebuilt_live_audio_room: 2.2.1; $version',
         tag: 'audio room',
         subTag: 'prebuilt',
       );
@@ -114,6 +114,7 @@ class _ZegoUIKitPrebuiltLiveAudioRoomState
       },
     );
     connectManager = ZegoLiveConnectManager(
+      config: widget.config,
       seatManager: seatManager,
       prebuiltController: widget.controller,
       innerText: widget.config.innerText,